refactor: controller -> controller + service 기존 컨트롤러의 역할을 서비스 클래스를 사용하여 MVC 패턴에 맞게 역할 분담함
